De-threaded MainController
